+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 * `smove` command.
+* `sdiff` command.
 
 ### Changed
 * added compat table badge to readme.