+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [2.2.0] - 2016-10-31
 
 ### Added
+* buffer operation aliases. (#234 @dpikt)
 * yarn lockfile.
 
 ### Changed
@@ -304,6 +305,7 @@ Bumping the version to v2 as there are fixes in this release that are breaking c
  * multi
  * exec
 
+[2.2.0]: https://github.com/stipsan/ioredis-mock/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/stipsan/ioredis-mock/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/stipsan/ioredis-mock/compare/v1.15.0...v2.0.0
 [1.15.0]: https://github.com/stipsan/ioredis-mock/compare/v1.14.0...v1.15.0